refactor: Rename display text map files to kebab-case
