+++ nextjs-app/constants/team-display-text-map.ts
